[android] - make used http client configurable
